Updated params

+++ app/server_manager/services/server_management/commands/create_site_command.py
@@ -30,28 +30,14 @@ class CreateSiteCommand(Command):
         return {"message": "Website server configuration created and applied successfully"}
 
     def create_fastcgi_params(self):
-        fastcgi_params = """
-fastcgi_param   QUERY_STRING        \\$query_string;
-fastcgi_param   REQUEST_METHOD      \\$request_method;
-fastcgi_param   CONTENT_TYPE        \\$content_type;
-fastcgi_param   CONTENT_LENGTH      \\$content_length;
-fastcgi_param   SCRIPT_FILENAME     \\$realpath_root\\$fastcgi_script_name;
-fastcgi_param   SCRIPT_NAME         \\$fastcgi_script_name;
-fastcgi_param   REQUEST_URI         \\$request_uri;
-fastcgi_param   DOCUMENT_URI        \\$document_uri;
-fastcgi_param   DOCUMENT_ROOT       \\$realpath_root;
-fastcgi_param   SERVER_PROTOCOL     \\$server_protocol;
-fastcgi_param   GATEWAY_INTERFACE   CGI/1.1;
-fastcgi_param   SERVER_SOFTWARE     nginx/\\$nginx_version;
-fastcgi_param   REMOTE_ADDR         \\$remote_addr;
-fastcgi_param   REMOTE_PORT         \\$remote_port;
-fastcgi_param   SERVER_ADDR         \\$server_addr;
-fastcgi_param   SERVER_PORT         \\$server_port;
-fastcgi_param   SERVER_NAME         \\$server_name;
-fastcgi_param   HTTPS               \\$https if_not_empty;
-fastcgi_param   REDIRECT_STATUS     200;
-fastcgi_param   HTTP_PROXY          \\"\\";
-"""
+        template_path = Path(__file__).parent.parent / 'templates' / 'nginx' / 'fastcgi_params_template.conf'
+        if not template_path.is_file():
+            raise FileNotFoundError(f"Template file not found: {template_path}")
+
+        with open(template_path, 'r') as template_file:
+            fastcgi_params = template_file.read()
+
+        # Write the content to the fastcgi_params file
         run_command(f'echo "{fastcgi_params.strip()}" | sudo tee /etc/nginx/fastcgi_params')
 
     def generate_dhparams(self):
